requirements.txt
Django==3.0.4
django-environ==0.4.5
django-material>=1.6.3,<2.0.0
django-material-admin>=1.6.44,<2.0.0
psycopg2-binary==2.8.4
django-crispy-bulma>=0.2,<1.0
django-crispy-forms>=1.9.0,<2.0.0
Pillow==7.0.0
django-spurl==0.6.6
django-admin-autocomplete-filter==0.4
boto3>=1.12.24,<2.0.0
django-storages>=1.9.1,<2.0.0
django-recaptcha>=2.0.6,<3.0.0
pycrypto>=2.6.1,<3.0.0
pyOpenSSL>=19.1.0,<20.0.0